add: `fetchClipboards` to global context provider
