+++ webui/src/context/ClipboardContext.tsx
@@ -7,6 +7,7 @@ type ClipboardContextType = {
   setClipboardList: React.Dispatch<
     React.SetStateAction<GetClipboardResponse[]>
   >;
+  fetchClipboards: () => Promise<void>;
 };
 
 const ClipboardContext = React.createContext<ClipboardContextType | null>(null);
@@ -26,8 +27,25 @@ export const ClipboardProvider: React.FC<ClipboardProviderProps> = ({
     GetClipboardResponse[]
   >([]);
 
+  const fetchClipboards = React.useCallback(async () => {
+    const response = await fetch("/api/clipboards");
+    if (response.status === 204) {
+      setClipboardList([]);
+      return;
+    }
+    if (response.status !== 200) {
+      console.error("Failed to get all clipboards");
+      return;
+    }
+    const text = await response.text();
+    const parsed: GetClipboardResponse[] = JSON.parse(text);
+    setClipboardList(parsed);
+  }, []);
+
   return (
-    <ClipboardContext.Provider value={{ clipboardList, setClipboardList }}>
+    <ClipboardContext.Provider
+      value={{ clipboardList, setClipboardList, fetchClipboards }}
+    >
       {children}
     </ClipboardContext.Provider>
   );
