+++ webui/src/pages/home/Home.tsx
@@ -3,7 +3,9 @@ function Home() {
     <>
       <h1>Home page</h1>
       <p>This page yet to be designed</p>
-      <p>For now feel free to checkout existing clipboards, or create your own!</p>
+      <p>
+        For now feel free to checkout existing clipboards, or create your own!
+      </p>
     </>
   );
 }
